Remove unused IndirectCommon functions
